requirement.txt
Flask
datetime
openai
mangum
werkzeug
mysql-connector-python